User administration DONE

+++ comdevelopmentboge/src/main/java/app/transactions/UserTypeChangeTransaction.java
@@ -38,7 +38,7 @@ public class UserTypeChangeTransaction {
             DbContext.getConnection().commit();
 
             MyAlert.showSuccess("Rola užívateľa " + user.getFullName() + " bola úspešne zmenená");
-            LogService.createLog("Zmena roly užívateľa " + user.getFullName());
+            LogService.createLog("Zmena roly užívatela " + user.getFullName());
         } catch (SQLException e) {
             user.setUserType(previousUserType);
             MyAlert.showSuccess("Rolu užívateľa " + user.getFullName() + " sa nepodarilo zmeniť");
@@ -52,23 +52,41 @@ public class UserTypeChangeTransaction {
     /**
      * Transakcia zmeny roly a projektov užívateľa
      * @param user - užívateľ, ktorému chceme zmeniť rolu a projekty
-     * @param projects - pridelené projekty
      * @throws DatabaseException
      * @throws SQLException
      */
-    public static void changeProjects(User user, List<Project> projects) throws DatabaseException, SQLException {
+    public static void changeProjects(User user, List<String> projectsToAdd, List<String> projectsToDelete) throws DatabaseException, SQLException {
         DbContext.getConnection().setAutoCommit(false);
         DbContext.getConnection().setTransactionIsolation(Connection.TRANSACTION_SERIALIZABLE);
         USERTYPE previousUserType = user.getUserTypeU();
         try {
             user.setUserType(USERTYPE.PROJECT_ADMIN);
+
+
+            for(String del : projectsToDelete) {
+                Administration a = AdministrationService.getAdministrationService().findAdministrationByProjectNum(del);
+                if (a != null) {
+                    a.delete();
+                }
+            }
+
+            for(String add : projectsToAdd) {
+                Project project = ProjectService.getProjectService().findProjectByNum(add);
+                Administration a = new Administration();
+                a.setProjectId(project.getId());
+                a.setUserId(user.getId());
+                a.insert();
+            }
+
             user.update();
 
-            for(Project project : projects) {
-                Administration a = AdministrationService.getAdministrationService().findAdministrationByProjectId(project.getId());
-                a.delete();
-            }
             DbContext.getConnection().commit();
+            if(previousUserType.equals(user.getUserTypeU())) {
+                LogService.createLog("Zmena projektov užívatela " + user.getFullName());
+            }
+            else {
+                LogService.createLog("Zmena roly a projektov užívatela " + user.getFullName());
+            }
             MyAlert.showSuccess("Rola a projekty užívateľa " + user.getFullName() + " boli úspešne zmenené");
         } catch (SQLException e) {
             user.setUserType(previousUserType);
@@ -81,73 +99,6 @@ public class UserTypeChangeTransaction {
 
     }
 
-    //TODO
-    public static void addProject(User user, String projectNum) throws SQLException {
-        DbContext.getConnection().setAutoCommit(false);
-        DbContext.getConnection().setTransactionIsolation(Connection.TRANSACTION_SERIALIZABLE);
-        try {
-            System.out.println("projectNum " + projectNum);
-            Project project = ProjectService.getProjectService().findProjectByNum(projectNum);
-            if(project == null) {
-                throw new MyException("Projekt sa nepodarilo vložiť");
-            }
-            System.out.println(project);
-            System.out.println("pr" + project.getProjectNumber());
-            Administration a = new Administration();
-            a.setUserId(user.getId());
-            System.out.println("UI" + user.getId());
-            System.out.println("PI" + project.getId());
-            a.setProjectId(project.getId());
-            a.insert();
-
-
-            DbContext.getConnection().commit();
-
-            DialogProjectsController.getInstance().reloadProjectsDelete();
-            DialogProjectsController.getInstance().reloadAddProjects();
-
-        } catch (SQLException e) {
-            MyAlert.showError("Projekt sa nepodarilo vložiť");
-            DbContext.getConnection().rollback();
-            throw e;
-        } catch (MyException e) {
-            MyAlert.showError(e.getMessage());
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            DbContext.getConnection().setAutoCommit(true);
-        }
-
-    }
-
-    public static void deleteProject(String projectNum) throws SQLException {
-        DbContext.getConnection().setAutoCommit(false);
-        DbContext.getConnection().setTransactionIsolation(Connection.TRANSACTION_SERIALIZABLE);
-        try {
-            System.out.println("projectNum " + projectNum);
-            Administration a = AdministrationService.getAdministrationService().findAdministrationByProjectNum(projectNum);
-            if(a == null) {
-                throw new MyException("Projekt sa nepodarilo vymazať");
-            }
-
-            a.delete();
-            DbContext.getConnection().commit();
-
-            DialogProjectsController.getInstance().reloadProjectsDelete();
-
-        } catch (SQLException e) {
-            MyAlert.showError("Projekt sa nepodarilo vymazať");
-            DbContext.getConnection().rollback();
-            throw e;
-        } catch (MyException e) {
-            MyAlert.showError(e.getMessage());
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            DbContext.getConnection().setAutoCommit(true);
-        }
-
-    }
